RuleEngine/src/main/java/com/zeotap/RuleEngine/service/RuleEngineService.java
package com.zeotap.RuleEngine.service;

import com.zeotap.RuleEngine.model.EligibilityRequest;
import com.zeotap.RuleEngine.model.Node;
import com.zeotap.RuleEngine.model.Rule;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Map;

@Service
public class RuleEngineService {

    public boolean evaluateEligibility(EligibilityRequest request) {
        // Implement your eligibility logic here
        // Example: check if age is greater than 18 and income is above a certain threshold
        return request.getAge() > 18 && request.getIncome() > 30000;
    }

    public Rule createRule(String ruleString) {
        return null;
    }

    public Node combineRules(List<String> rules) {
        return null;
    }

    public boolean evaluateRule(Rule ast, Map<String, Object> data) {
        return false;
    }

    public Rule updateRule(Long id, Rule rule) {
        return rule;
    }

    public boolean deleteRule(Long id) {
        return false;
    }

    public List<Rule> getAllRules() {
        return List.of();
    }

    public Rule getRuleById(Long id) {
        return null;
    }

    public boolean evaluateRule(Node combinedAst, Map<String, Object> data) {
        return false;
    }
}
